Fixed bugs for indexing when word does not exist in index
