+++ backend.py
@@ -4,24 +4,30 @@ from datetime import datetime
 import logic
 import json
 from collections import defaultdict
+import parser
 app = Flask(__name__)
 
 @app.route('/', methods =["GET", "POST"])
 def homepage():
+    sentence_List = []
+    language_list = []
+    part_of_speech_list = []
+    error = ""
     # Get the indexing in memory so that we have it until the application is closed
     dictionary = logic.GetIndexing()
-    if request.method == "GET":
-        sentence_List = ""
-        error = ""
+    if(len(dictionary) == 0):
+        error = "Indexing file is not present"
     if request.method == "POST":
         language_selected = request.form.get('language')
         Lang_ID = logic.GetLanguageId(language_selected)
         part_of_speech_selected = request.form.get('partOfSpeech')
         word_selected = (request.form.get('word') + '/' + request.form.get('partOfSpeech')).lower()
+        if(not logic.isCorpusLoaded(language_selected + "_corpus")):
+            error = "Corpus is not loaded into the database"
         # ignore first and last characters i.e. '[' and ']' to get the list of line ids as a string like "1,3,6,7,...""
-        line_ids= str(dictionary[word_selected])[1:][:-1]
-        sentence_List = logic.SQLQuery(f"select Line_Text from {language_selected}_corpus where Line_id in ({line_ids})")
-        error = ""
+        if word_selected in dictionary:
+            line_ids= str(dictionary[word_selected])[1:][:-1]
+            sentence_List = logic.SQLQuery(f"select Line_Text from {language_selected}_corpus where Line_id in ({line_ids})")
         if len(sentence_List) == 0:
             error = "Error: Word not in corpus"
     language_list = logic.SQLQuery("SELECT Lang_Desc FROM Lang_Ref;")
