+++ parser.py
@@ -13,25 +13,6 @@ from collections import defaultdict
 import json
 nltk.download('punkt')
 
-def CreateIndexing(sentence, line_id, dictionary):
-    word_list = sentence.split()
-    # number of parts formed for each tagged word in the corpus(word/pos)
-    tagged_word_parts = 2   
-    for word in word_list:
-        # skip untagged words
-        if(len(word.split('/')) != tagged_word_parts):
-            continue
-        subtag = (word.split('/')[1]).upper()
-        # ignore -TL suffix in subtags since they are just to indicate that the word occurs in title
-        subtag = subtag.split('-TL')[0]
-        if(subtag in logic.english_pos_mapping):
-            generalized_pos = logic.english_pos_mapping[subtag]
-        else:
-            generalized_pos = subtag        
-        word = (word.split('/')[0] + "/" + generalized_pos).lower()
-        dictionary[word].append(line_id)        
-    return dictionary
-
 # Example path: 'E:\\SLU\\Sem1\\Principles of SD\\Project Material\\brown'
 path = input("Enter path of the corpus directory: ")
 language = input("Enter language: ")
@@ -59,12 +40,13 @@ for file_name in file_name_list:
         for sentence in sentence_list:
             # form the values string required in insert query
             values += "(" + language_id + ", " + str(document_id) + ", " + "'" + file_name + "', " + "\"" + sentence + "\", " + "'" + str(datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')) + "'),"
-            dictionary = CreateIndexing(sentence, line_id, dictionary)
+            dictionary = logic.CreateIndexing(sentence, line_id, dictionary)
             line_id += 1
 # remove last comma from values string    
 values = values[:-1]
 corpus_table = language.lower() + "_corpus"
 query = "insert into " + corpus_table + " (Lang_ID, Doc_ID, Doc_Name, Line_Text, Last_Update) values " + values
-logic.SQLInsertQuery(query)
+if(not logic.isCorpusLoaded(corpus_table)):
+    logic.SQLInsertQuery(query)
 document_id += 1
 logic.StoreIndexing(json.dumps(dictionary))
